requirements.txt
aiohttp==3.9.5
python-telegram-bot==21.3
feedparser==6.0.11
apscheduler==3.10.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
